+++ client/web-admin-2/src/services/utils.js
@@ -1,0 +1,6 @@
+
+export const pick = (obj, keys) => {
+  const out = {}
+  keys.forEach((key) => out[key] = obj[key])
+  return out
+}
